bugfix: add file checking before removing it

+++ paddlenlp/transformers/model_utils.py
@@ -878,7 +878,12 @@ class PretrainedModel(Layer, GenerationMixin):
             if weight_file_path != new_weight_file_path:
                 # copy the weight file
                 shutil.copyfile(weight_file_path, new_weight_file_path)
-                os.remove(weight_file_path)
+
+                # Raise by QA(junjun): it's strange that `shutil.copyfile` will remove `weight_file_path` file.
+                # so do checking before removin it,
+                if os.path.exist(weight_file_path):
+                    os.remove(weight_file_path)
+
                 # shutil.rmtree(weight_file_path)
                 weight_file_path = new_weight_file_path
